.github/workflows: add node-packages.nix check
This workflow intends to enforce that PRs that add packages to
node-packages.json should not also update node-packages.nix, because
of the high chance of conflicts. The plan is to have the node packages
maintainers separately do batch updates to node-packages.nix on a
regular interval (hopefully weekly).

The maintainers should add [nodepackages ci skip] to their commit
message to bypass this check.